Fixes to simultaneous ADC and FFT

+++ src/boot/ksdk1.1.0/FFT.h
@@ -3,6 +3,6 @@
 
 void FFT(int complex * x, uint8_t N);
 void FFTN(int complex * x, uint8_t n);
-void bitReverse(int * x, uint8_t N);
+void bitReverse(uint16_t * x, uint8_t N);
 void applyWindow32(int * samples);
 void octaves(int complex * x, uint16_t * output, uint8_t N);
